Add files via upload
In the world of data science and machine learning, every project is an exciting adventure waiting to be embarked upon. Recently, I delved into the realm of predictive modeling by creating a random dataset and putting my knowledge to the test using Kaggle notebooks.

The first step was to gather and prepare the data. The dataset was random and generated using Chat GPT.

With a neatly organized dataset in hand, I set out to apply Linear Regression models. This classic machine-learning technique is a fundamental building block for predictive modeling. It allowed me to establish relationships between different variables in the data and make predictions based on those relationships.

The variables of interest were Salary & Years of Experience

To evaluate the performance of my models, I employed three key metrics: Mean Squared Error (MSE), R-squared (R2) Score, and Mean Absolute Error (MAE). These metrics provided valuable insights into the accuracy and reliability of the models. The lower the MSE and MAE and the higher the R2 score, the better the model's predictions.

Results from the linear regression model show that an increase in years of experience is positively associated with an increase in salary


The Link to the Kaggle notebook: https://lnkd.in/d8wtJZ8y